[tools] Rename find_untranslated_strings.py to strings_utils.py
