+++ tools/python/strings_utils.py
